Separate github api calls
