+++ apps/charterafrica/src/lib/ecosystem/index.js
@@ -35,6 +35,7 @@ export async function updateList() {
       level: "info",
     });
     const data = await airtable.data(config);
+    // console.log()
     await prepareContributors(data, config);
     await prepareTools(data, config);
     await prepareOrganisations(data, config);
